LuaTool/requirements.txt
discord_webhook
requests
colorama
plyer
discord.py
selenium
beautifulsoup4
Pillow
PyAutoGUI
pyperclip
selenium
aiohttp
pypresence
pyinstaller
dhooks
pystyle
beautifulsoup4
